Moved test into separate file
