+++ tests/test_various.rs
@@ -3,6 +3,8 @@ use pyo3::types::IntoPyDict;
 use pyo3::types::{PyDict, PyTuple};
 use pyo3::{py_run, wrap_pyfunction, PyCell};
 
+use std::fmt;
+
 mod common;
 
 #[pyclass]
@@ -167,4 +169,42 @@ fn test_pickle() {
         assert inst2.__dict__ == {'a': 1}
     "#
     );
+}
+
+/// Testing https://github.com/PyO3/pyo3/issues/1106. A result type that
+/// implements `From<MyError> for PyErr` should be automatically converted
+/// when using `#[pyfunction]`.
+///
+/// This only makes sure that valid `Result` types do work. For an invalid
+/// enum type, see `ui/invalid_result_conversion.py`.
+#[derive(Debug)]
+struct MyError {
+    pub descr: String,
+}
+
+impl fmt::Display for MyError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "My error message: {}", self.descr)
+    }
+}
+
+/// Important for the automatic conversion to `PyErr`.
+impl From<MyError> for PyErr {
+    fn from(err: MyError) -> pyo3::PyErr {
+        pyo3::exceptions::PyOSError::py_err(err.to_string())
+    }
+}
+
+#[pyfunction]
+fn result_conversion_function() -> Result<(), MyError> {
+    Err(MyError {
+        descr: "something went wrong".to_string(),
+    })
+}
+
+#[test]
+fn test_result_conversion() {
+    let gil = Python::acquire_gil();
+    let py = gil.python();
+    wrap_pyfunction!(result_conversion_function)(py);
 }
